Multi Agent Systems/requirements.txt
crewai==0.134.0
crewai-tools==0.49.0
duckduckgo_search==8.0.5
pydantic==2.11.7
PyYAML==6.0.2
pyyaml_env_tag==1.1
langchain==0.3.26
langchain-community==0.3.27
python-dotenv>=0.19.0
openai>=1.0.0 
